requirements.txt
fastapi
uvicorn
numpy
pillow
aiofiles
python-multipart
